requirements.txt
# RSS Feed Processing
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI and ML
openai>=1.0.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Web Scraping
selenium>=4.10.0
playwright>=1.40.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# GitHub Integration
pygithub>=1.59.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Logging and Monitoring
loguru>=0.7.0
prometheus-client>=0.17.0

# Async Support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0